Tidied up CURUniqueRequests.h
