+++ adagucserverEC/CUniqueRequests/CURConstants.h
@@ -1,2 +1,4 @@
+#include <map>
+#include "CTString.h"
 #define CCUniqueRequests_MAX_DIMS 255
 typedef std::map<int, CT::string *>::iterator it_type_dimvalindex;
